+++ tests/web/test_about.py
@@ -18,5 +18,4 @@ async def test_about_page_shows_app_info(user: User) -> None:
     await user.should_see(info.version)
     await user.should_see(info.description)
     await user.should_see(info.author_email)
-    await user.should_see("GitHub")
-    await user.should_see("Documentation")
+    await user.should_see("Project Url")